+++ lexer.h
@@ -30,7 +30,7 @@ typedef struct Token {
     size_t line; /* line number */
     enum TokenType type;
     union {
-        int32_t num;
+        int64_t num;
         struct {
             const char *ptr;
             size_t len;
