+++ object.h
@@ -59,7 +59,7 @@ struct Object {
     bool eval;
     union {
         struct StringSlice str;
-        int32_t num;
+        int64_t num;
         struct List list;
         Builtin builtin;
         struct Function function;
@@ -70,7 +70,7 @@ Object *object_new_generic(void);
 Object *object_list_new(Object *car, Object *cdr);
 Object *object_string_slice_new(const char *s, size_t len);
 Object *object_ident_new(const char *s, size_t len);
-Object *object_num_new(int32_t num);
+Object *object_num_new(int64_t num);
 Object *object_nil_new(void);
 Object *object_builtin_new(Builtin f);
 const char *object_type_as_string(enum ObjectKind k);
